separate config load with support for package.json
